Initial refactor moving uci to its own class
